requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.0.0
et-xmlfile==1.0.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
jdcal==1.4.1
oauthlib==3.1.0
openpyxl==3.0.6
ordereddict==1.1
packaging==20.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-decouple==3.4
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
WooCommerce==2.1.1

